requirements-new-3_7.txt
anyio==3.6.2
boto3==1.26.45
botocore==1.29.45
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.88.0
filelock==3.9.0
huggingface-hub==0.11.1
idna==3.4
jmespath==1.0.1
joblib==1.2.0
nltk==3.8.1
numpy==1.21.4
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.3.5
pandas-stubs==1.2.0.62
Pillow==9.4.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
threadpoolctl==3.1.0
tokenizers==0.13.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.64.1
transformers==4.25.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0